Rework test setups so that they lead to different results under 2.14.0 vs 2.15.2 `git machete discover`
