+++ scripts/run-ui-tests
@@ -9,6 +9,8 @@ source "$self_dir"/utils.sh
 intellij_version=${1-} # can be skipped; the same IntelliJ version that the plugin has been built against will be assumed
 
 ide_log=build/idea-sandbox/system/log/idea.log
+# idea.log is not automatically cleared before each `runIde(forUiTests)`
+# and we don't want to print the potential logs from the previous runs.
 ide_log_last_line=$(cat $ide_log 2>/dev/null | wc -l || true)  # will be 0 if $ide_log doesn't exist yet
 
 function ide_logs_from_this_run() {
